Undo change variable name for nameTitleGroup attribute in name elements
The nameTitleGroup attribute is meant to link titleInfo and name elements. The data entered
is supposed to be the same in order to link the two elements. When entering data in the
spreadsheet, users may not neccessarily want ti1:titleInfo to link to na1:name, but rather
link ti2:titleInfo to na1:name. This would happen when there is the transcribed title for
a resource to be entered into ti1:titleInfo and a uniform name-title to be entered in ti2:titleInfo
and na1:name. In order to link ti2:titleInfo and na1:name, the user would have to have the
ability to enter in a value for ti2:nameTitleGroup and na1:nameTitleGroup.

This change allows users to link the correct fields. But it also introduces the possibility
of users unintentionally not linking fields that they meant to link.